Move formatRestrictions tests
