+++ apps/judicial-system/backend/src/app/formatters/formatters.spec.ts
@@ -4,7 +4,6 @@ import {
   CaseCustodyRestrictions,
   CaseGender,
 } from '@island.is/judicial-system/types'
-import { formatRestrictions } from '@island.is/judicial-system/formatters'
 import {
   formatConclusion,
   formatCourtCaseNumber,
@@ -196,82 +195,6 @@ describe('formatConclusion', () => {
     // Assert
     expect(res).toBe(
       'Kærði, Glanni Glæpur 010101-0000 skal sæta gæsluvarðhaldi, þó ekki lengur en til þriðjudagsins 22. desember 2020 kl. 11:23. Kærði skal sæta einangrun meðan á gæsluvarðhaldi stendur.',
-    )
-  })
-})
-
-describe('formatRestrictions', () => {
-  test('should return formatted restrictions for no restrictions', () => {
-    // Arrange
-    const custodyRestrictions = []
-
-    // Act
-    const res = formatRestrictions(custodyRestrictions)
-
-    // Assert
-    expect(res).toBe('Sækjandi tekur fram að gæsluvarðhaldið sé án takmarkana.')
-  })
-
-  test('should return formatted restrictions for isolation only', () => {
-    // Arrange
-    const custodyRestrictions = [CaseCustodyRestrictions.ISOLATION]
-
-    // Act
-    const res = formatRestrictions(custodyRestrictions)
-
-    // Assert
-    expect(res).toBe(
-      'Sækjandi tekur fram að kærði skuli sæta einangrun meðan á gæsluvarðhaldi stendur.',
-    )
-  })
-
-  test('should return formatted restrictions for isolation and one other restriction', () => {
-    // Arrange
-    const custodyRestrictions = [
-      CaseCustodyRestrictions.ISOLATION,
-      CaseCustodyRestrictions.MEDIA,
-    ]
-
-    // Act
-    const res = formatRestrictions(custodyRestrictions)
-
-    // Assert
-    expect(res).toBe(
-      'Sækjandi tekur fram að kærði skuli sæta einangrun meðan á gæsluvarðhaldi stendur og að gæsluvarðhaldið verði með fjölmiðlabanni skv. 99. gr. laga nr. 88/2008.',
-    )
-  })
-
-  test('should return formatted restrictions for all but isolation', () => {
-    // Arrange
-    const custodyRestrictions = [
-      CaseCustodyRestrictions.COMMUNICATION,
-      CaseCustodyRestrictions.MEDIA,
-      CaseCustodyRestrictions.VISITAION,
-    ]
-
-    // Act
-    const res = formatRestrictions(custodyRestrictions)
-
-    // Assert
-    expect(res).toBe(
-      'Sækjandi tekur fram að gæsluvarðhaldið verði með bréfaskoðun og símabanni, fjölmiðlabanni og heimsóknarbanni skv. 99. gr. laga nr. 88/2008.',
-    )
-  })
-
-  test('should order non-isolation restrictions', () => {
-    // Arrange
-    const custodyRestrictions = [
-      CaseCustodyRestrictions.MEDIA,
-      CaseCustodyRestrictions.VISITAION,
-      CaseCustodyRestrictions.COMMUNICATION,
-    ]
-
-    // Act
-    const res = formatRestrictions(custodyRestrictions)
-
-    // Assert
-    expect(res).toBe(
-      'Sækjandi tekur fram að gæsluvarðhaldið verði með bréfaskoðun og símabanni, fjölmiðlabanni og heimsóknarbanni skv. 99. gr. laga nr. 88/2008.',
     )
   })
 })
